"Fattening" of islands now works.
A 3d matrix now describes which cells should be opened when clicking on
a 0-field. (By iterating through the whole board)

Therefore I am now able to start work on the more efficient lookup
table solution for 0-fields. (By just using a look-up table for the
fields to be opened. This only iterates through the fields that should
open. Much more efficient for larger boards!)

For example, if the field is only 10 cells, but the board 10 000 cells,
the look-up solution will be 1 000 times more efficient! This is
very important for large boards.